@@ -11,41 +11,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.14",
-    "@angular/common": "~10.0.9",
-    "@angular/compiler": "~10.0.9",
-    "@angular/core": "~10.0.9",
-    "@angular/forms": "~10.0.9",
-    "@angular/platform-browser": "~10.0.9",
-    "@angular/platform-browser-dynamic": "~10.0.9",
-    "@angular/router": "~10.0.9",
+    "@angular/animations": "~12.2.7",
+    "@angular/common": "~12.2.7",
+    "@angular/compiler": "12.2.7",
+    "@angular/core": "~12.2.7",
+    "@angular/forms": "~12.2.7",
+    "@angular/platform-browser": "~12.2.7",
+    "@angular/platform-browser-dynamic": "~12.2.7",
+    "@angular/router": "~12.2.7",
     "@auth0/angular-jwt": "^5.0.2",
-    "@fortawesome/angular-fontawesome": "^0.7.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
-    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@babel/core": "^7.15.5",
+    "@fortawesome/angular-fontawesome": "^0.9.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@ngrx/effects": "^12.4.0",
+    "@ngrx/store": "^12.4.0",
+    "@ngrx/store-devtools": "^12.4.0",
+    "fortawesome": "0.0.1-security",
+    "ngrx-store-localstorage": "^12.0.1",
     "ngx-toastr": "^13.2.0",
     "rxjs": "~6.5.5",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.6",
-    "@angular/cli": "~10.0.6",
-    "@angular/compiler-cli": "~10.0.9",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "^12.2.7",
+    "@angular/cli": "12.2.7",
+    "@angular/compiler-cli": "12.2.7",
     "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "@types/jasminewd2": "2.0.10",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.8",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "tslint": "~6.1.3",
+    "typescript": "~4.3.2"
   }
 }
