@@ -5,14 +5,13 @@ import { GuideGuard } from './core/gards/guide.guard';
 
 
 
-
 const routes: Route[] = [
-  {path:'',pathMatch:'full',redirectTo:'home'},
-  {path:'sample',loadChildren:'./sample/sample.module#SampleModule'},  
-  {path:'user',loadChildren:'./user/user.module#UserModule'}, 
-  {path:'tour',loadChildren:'./tour/tour.module#TourModule'},
-  {path:'category',loadChildren:'./category/category.module#CategoryModule',canActivateChild:[AuthGuardGuard,GuideGuard]},
-  {path:'buy',loadChildren:'./buy/buy.module#BuyModule',canActivateChild:[AuthGuardGuard]},
+  {path:'',pathMatch:'full',redirectTo:'/home'},
+  {path:'sample',loadChildren:() => import('./sample/sample.module').then(m => m.SampleModule)},  
+  {path:'user',loadChildren: () => import('./user/user.module').then(m => m.UserModule) }, 
+  {path:'tour',loadChildren:() => import('./tour/tour.module').then(m => m.TourModule)},
+  {path:'category',loadChildren:() => import('./category/category.module').then(m => m.CategoryModule),canActivateChild:[AuthGuardGuard,GuideGuard]},
+  {path:'buy',loadChildren:() => import('./buy/buy.module').then(m => m.BuyModule),canActivateChild:[AuthGuardGuard]},
   
 
   { path: '**', redirectTo: '' } 
