@@ -10,13 +10,12 @@ import { CategoryService } from '../../core/service/category.service'
 })
 export class CreateCategoryComponent implements OnInit {
   form: FormGroup;
-
-  constructor(private fb: FormBuilder, private categoryService: CategoryService, private router: Router) { }
+  constructor(private fb: FormBuilder, private categoryService: CategoryService, private router: Router) {}
 
   ngOnInit(): void {
     this.form = this.fb.group({
       name: ['', Validators.required],
-    })
+    });    
   }
 
   get f() {
