@@ -1,21 +1,33 @@
 import { Injectable } from '@angular/core';
-import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http'
+import {
+  HttpEvent,
+  HttpHandler,
+  HttpInterceptor,
+  HttpRequest,
+} from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { Store } from '@ngrx/store';
+import { auth } from '../../+store';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class JwtInterceptorService implements HttpInterceptor {
+  token: string;
+  
+  constructor(private store: Store) {
+    this.store.select(auth.getToken).subscribe((t) => (this.token = t));
+  }
 
-  constructor() { }
-
-  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+  intercept(
+    req: HttpRequest<any>,
+    next: HttpHandler
+  ): Observable<HttpEvent<any>> {
     const cloneReq = req.clone({
-      setHeaders:
-      {
-        Authorization: `Bearer ${localStorage.getItem('token')}`
-      }
-    })
+      setHeaders: {
+        Authorization: `Bearer ${this.token}`,
+      },
+    });
     return next.handle(cloneReq);
   }
 }
