@@ -13,24 +13,34 @@ import {SampleModule} from '../sample/sample.module';
 import { AuthGuardGuard } from '../core/gards/auth-guard.guard';
 import {GuideGuard} from '../core/gards/guide.guard';
 import { UpdateComponent } from './update/update.component';
-
-
+import {StoreModule} from '@ngrx/store';
+import {globalReducer} from '../+store/global/reducer';
+import { EffectsModule } from '@ngrx/effects';
+import {GlobalEffects} from '../+store/global/effects';
+import {TourEffects} from '../+store/tour/effects';
+import {tourReducer} from '../+store/tour/reducer';
+import { TourListComponent } from './guide-tours/tour-list/tour-list.component';
+import {tourMetaReducer} from '../+store/meta-reducer';
 
 
 @NgModule({
-  declarations: [CreateComponent, TourCardComponent, DetailsComponent, RandomComponent, TourDescriptionComponent, UpdateComponent],
+  declarations: [CreateComponent, TourCardComponent, DetailsComponent, RandomComponent, TourDescriptionComponent, UpdateComponent, TourListComponent],
   imports: [
     CommonModule,
     FontAwesomeModule,
     ReactiveFormsModule,
     SampleModule,
     RouterModule.forChild([
-      { path: 'create', component: CreateComponent,canActivate:[AuthGuardGuard,GuideGuard],  },
+      { path: 'create', pathMatch: 'full', component: CreateComponent,canActivate:[AuthGuardGuard,GuideGuard]},
       { path: 'random', component: RandomComponent },
       {path:'tour-card', component:TourCardComponent},
-      { path: 'description/:id', component: TourDescriptionComponent,canActivate:[AuthGuardGuard] },
-      {path:'update/:id',component: UpdateComponent, canActivate:[AuthGuardGuard,GuideGuard]}      
-    ])
+      {path: 'description/:id', component: TourDescriptionComponent,canActivate:[AuthGuardGuard] },
+      {path:'update/:id',component: UpdateComponent,canActivate:[AuthGuardGuard,GuideGuard]}, 
+      {path:'list',component:TourListComponent,canActivate:[AuthGuardGuard,GuideGuard]}     
+    ]),
+    StoreModule.forFeature('global',globalReducer),  
+    StoreModule.forFeature('tour',tourReducer,{metaReducers:tourMetaReducer}),  
+    EffectsModule.forFeature([GlobalEffects,TourEffects])
   ],
   exports: [
     CreateComponent,
